refactor(ast): restructure the module
