+++ packages/utilities/ast/src/ast-is.ts
@@ -1,3 +1,4 @@
+import type { TSESTree } from "@typescript-eslint/types";
 import { or } from "@eslint-react/eff";
 import { AST_NODE_TYPES as T } from "@typescript-eslint/types";
 import { ASTUtils } from "@typescript-eslint/utils";
