+++ packages/utilities/ast/src/ast-node-equal.ts
