+++ packages/utilities/ast/src/ast-process-env-node-env.ts
@@ -1,8 +1,24 @@
 import type { _ } from "@eslint-react/eff";
 import type { TSESTree } from "@typescript-eslint/types";
 import { AST_NODE_TYPES as T } from "@typescript-eslint/types";
-import { isLiteral } from "./is-literal";
-import { isProcessEnvNodeEnv } from "./is-process-env-node-env";
+import { isLiteral } from "./ast-literal-is";
+
+/**
+ * Check if the given node is a member expression that accesses `process.env.NODE_ENV`
+ * @param node The AST node
+ * @returns True if the node is a member expression that accesses `process.env.NODE_ENV`, false otherwise
+ */
+export function isProcessEnvNodeEnv(node: TSESTree.Node | null | _): node is TSESTree.MemberExpression {
+  return node != null
+    && node.type === T.MemberExpression
+    && node.object.type === T.MemberExpression
+    && node.object.object.type === T.Identifier
+    && node.object.object.name === "process"
+    && node.object.property.type === T.Identifier
+    && node.object.property.name === "env"
+    && node.property.type === T.Identifier
+    && node.property.name === "NODE_ENV";
+}
 
 /**
  * Check if the given node is a binary expression that compares `process.env.NODE_ENV` with a string literal
