+++ packages/utilities/ast/src/ast-property-name.ts
@@ -1,12 +1,12 @@
 import { _ } from "@eslint-react/eff";
 import { AST_NODE_TYPES as T, type TSESTree } from "@typescript-eslint/types";
 
-import { getEcmaExpression } from "./get-ecma-expression";
-import { isTypeExpression } from "./is";
+import { getJSExpression } from "./ast-expression";
+import { isTypeExpression } from "./ast-is";
 
 export function getPropertyName(node: TSESTree.Node): string | _ {
   if (isTypeExpression(node)) {
-    return getPropertyName(getEcmaExpression(node));
+    return getPropertyName(getJSExpression(node));
   }
   if (node.type === T.Identifier || node.type === T.PrivateIdentifier) {
     return node.name;
