--- packages/utilities/ast/src/get-literal-value-type.ts
@@ -1,5 +1,0 @@
-export function getLiteralValueType(input: bigint | boolean | null | number | string | symbol) {
-  // eslint-disable-next-line local/prefer-eqeq-nullish-comparison
-  if (input === null) return "null";
-  return typeof input;
-}
